renamed BenVisual to BenView and separated it
