+++ Makefile
@@ -1,13 +1,13 @@
 all: driver
 
-driver: ben.o driver.o benvisual.o
-	g++ -o driver ben.o driver.o benvisual.o -lncurses -lpanel -pthread
+driver: ben.o benview.o driver.o
+	g++ -o driver ben.o benview.o driver.o -lncurses -lpanel -pthread
 
-ben.o: ben.h ben.cpp
+ben.o: ben.h benview.h ben.cpp
 	g++ -c ben.cpp
 
-benvisual.o: ben.h benvisual.cpp
-	g++ -c benvisual.cpp
+benview.o: benview.h benview.cpp
+	g++ -c benview.cpp
 
 driver.o: driver.cpp ben.h
 	g++ -c driver.cpp
